Document user private groups and primary GIDs
Add some explicit documentation of user private groups, since that
wasn't fleshed out very well.  Document that they may not be in use
for OpenID Connect deployments.  Document the new primary GID field
and mention that it's required for the Notebook Aspect.  Discuss how
UID and GID are used when spawning lab pods.