Update for recent Gafaelfawr changes
Document that using `GafaelfawrIngress` is now required, not just
recommended. Document the new service-to-service authentication
approach when the user should not be able to directly access the
underlying service. Document the move of ingress-specific routes
to private routes and the use of the cluster-internal address to
Gafaelfawr, along with the implications for vCenter not being
supported.

Add some documentation of how the state attribute in cookies is
handled during authentication.

Remove baseUrl from the example `GafaelfawrIngress` resource, since
this is no longer needed.

Add a section on metrics support using Sasquatch.